move ThunderModule to separate file
